v0.25
Fully functioning dac now.
a) UI to come.
b) Can edit variables in the python script for now. (line 19 or so).
c) Pulse trians only for now.
d) Tested to 5K only for now.

Use: Load the pyStim.ino to a teensy 3.5 or 3.6
Keep it connected and edit the com port for your board. Then trigger the python script when you want a tain. The teensy led will blink when it is pulsing.